Move gui components to own subpackage
