+++ src/arho_feature_template/gui/forms/feature_attribute_form.ui
