+++ src/arho_feature_template/gui/panels/template_library_panel.ui
